requirements.txt
# 核心依赖
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.5.0
click>=8.1.0
rich>=13.7.0

# 数据处理
orjson>=3.9.0

# 配置
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 开发依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# 类型存根
types-requests>=2.31.0
types-pyyaml>=6.0.0
